requirements.txt
aio-pika==9.4.3
aiormq==6.8.1
astroid==3.3.5
black==24.10.0
click==8.1.7
dill==0.3.9
dnspython==2.7.0
idna==3.10
isort==5.13.2
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pamqp==3.3.0
pathspec==0.12.1
pika==1.3.2
platformdirs==4.3.6
propcache==0.2.0
pylint==3.3.1
pymongo==4.10.1
python-dotenv==1.0.1
tomlkit==0.13.2
yarl==1.17.0
